requirements.txt
pynput==1.8.0
python-dotenv==1.0.1
